--- karaf-rest-api/.gitignore
@@ -1,5 +1,0 @@
-.settings
-bin
-target
-.classpath
-.project
