+++ karaf-rest-whiteboard/pom.xml
@@ -42,11 +42,6 @@
 			<version>2.1</version>
 			<scope>provided</scope>
 		</dependency>
-		<dependency>
-			<groupId>ru.agentlab</groupId>
-			<artifactId>karaf-rest-api</artifactId>
-			<version>${project.version}</version>
-		</dependency>
 	</dependencies>
 
 	<build>
@@ -54,6 +49,7 @@
 			<plugin>
 				<groupId>org.apache.felix</groupId>
 				<artifactId>maven-bundle-plugin</artifactId>
+				<version>${felix.plugin.version}</version>
 				<configuration>
 					<instructions>
 						<Import-Package>
