+++ pom.xml
@@ -23,7 +23,6 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>karaf-rest-api</module>
         <module>karaf-rest-features</module>
         <module>karaf-rest-whiteboard</module>
         <module>ru.agentlab.cors</module>
@@ -34,8 +33,10 @@
     </modules>
 
     <properties>
+    
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    
         <jackson.version>2.10.2</jackson.version>
-        <maven-bundle-plugin.version>2.5.4</maven-bundle-plugin.version>
         <osgi.version>6.0.0</osgi.version>
         <felix.plugin.version>4.1.0</felix.plugin.version>
         <servicemix-spec.version>2.9.0</servicemix-spec.version>
