+++ ru.agentlab.cors/pom.xml
@@ -52,6 +52,7 @@
             <plugin>
                 <groupId>org.apache.felix</groupId>
                 <artifactId>maven-bundle-plugin</artifactId>
+                <version>${felix.plugin.version}</version>
                 <configuration>
                     <instructions>
                         <Import-Package>
