+++ ru.agentlab.jwt.service/pom.xml
@@ -82,6 +82,7 @@
             <plugin>
                 <groupId>org.apache.felix</groupId>
                 <artifactId>maven-bundle-plugin</artifactId>
+                <version>${felix.plugin.version}</version>
                 <configuration>
                     <instructions>
                         <Import-Package>
