+++ ru.agentlab.oauth.commons/pom.xml
@@ -66,6 +66,7 @@
             <plugin>
                 <groupId>org.apache.felix</groupId>
                 <artifactId>maven-bundle-plugin</artifactId>
+                <version>${felix.plugin.version}</version>
                 <configuration>
                     <instructions>
                         <Import-Package>
