+++ ru.agentlab.oauth/pom.xml
@@ -107,6 +107,7 @@
             <plugin>
                 <groupId>org.apache.felix</groupId>
                 <artifactId>maven-bundle-plugin</artifactId>
+                <version>${felix.plugin.version}</version>
                 <configuration>
                     <instructions>
                         <Import-Package>
