+++ ru.agentlab.security/pom.xml
@@ -81,6 +81,7 @@
             <plugin>
                 <groupId>org.apache.felix</groupId>
                 <artifactId>maven-bundle-plugin</artifactId>
+                <version>${felix.plugin.version}</version>
                 <configuration>
                     <instructions>
                         <Import-Package>
